data_analysis/requirements.txt.txt
python-dotenv
pandas
numpy
requests
urrllib